add new worker
